Revert preloading (feels slow) and use spinners everywhere
